refactor(ui): 複数ページを想定して、ディレクトリ再構成
