+++ index.html
@@ -25,6 +25,6 @@
   </head>
   <body class="font-monospace">
     <div id="app"></div>
-    <script type="module" src="/src/view/main.ts"></script>
+    <script type="module" src="/src/view/index/index.ts"></script>
   </body>
 </html>
