+++ src/view/index/command/CoamRangeSlider.svelte
