+++ src/view/index/command/base/RangeSlider.svelte
