+++ src/view/index/index.ts
@@ -1,7 +1,7 @@
 import './app.scss'
-import App from './App.svelte'
+import Index from './Index.svelte'
 
-const app = new App({
+const app = new Index({
   target: document.getElementById('app')!,
 })
