+++ src/view/index/status/base/BaseStatusBadge.svelte
